requirements.txt
strands-agents 
strands-agents-tools
bedrock_agentcore
Appium-Python-Client
pyperclip
pyautogui
Pillow
pyppeteer